Refactors to support package importing
