+++ cmd/main_test.go
@@ -1,0 +1,4 @@
+// gcov2lcov - convert golang coverage files to the lcov format.
+// (c) 2019 Jan Delgado
+package main
+
